+++ myst_nb/__init__.py
@@ -27,6 +27,6 @@ def glue(name: str, variable, display: bool = True) -> None:
         state of the object at glue-time.
     """
     # we import this locally, so IPython is not automatically imported
-    from myst_nb.nb_glue import glue
+    from myst_nb.glue import glue
 
     return glue(name, variable, display)
