+++ pyproject.toml
@@ -51,14 +51,14 @@ Homepage = "https://github.com/executablebooks/myst-nb"
 Documentation = "https://myst-nb.readthedocs.io"
 
 [project.entry-points."myst_nb.renderers"]
-default = "myst_nb.render:NbElementRenderer"
+default = "myst_nb.core.render:NbElementRenderer"
 
 [project.entry-points."pygments.lexers"]
-myst-ansi = "myst_nb.lexers:AnsiColorLexer"
-ipythontb = "myst_nb.lexers:IPythonTracebackLexer"
+myst-ansi = "myst_nb.core.lexers:AnsiColorLexer"
+ipythontb = "myst_nb.core.lexers:IPythonTracebackLexer"
 
 [project.entry-points."jcache.readers"]
-myst_nb_md = "myst_nb.read:myst_nb_reader_plugin"
+myst_nb_md = "myst_nb.core.read:myst_nb_reader_plugin"
 
 [project.optional-dependencies]
 code_style = ["pre-commit~=2.12"]
