+++ tests/test_glue.py
@@ -3,7 +3,7 @@ from IPython.core.interactiveshell import InteractiveShell
 import nbformat
 import pytest
 
-from myst_nb.nb_glue import extract_glue_data, glue
+from myst_nb.glue import extract_glue_data, glue
 
 
 class MockDisplayPublisher(DisplayPublisher):
